requirements.txt
customerio==5.4.0
python-dotenv==1.0.0